% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weco_rules.R
\name{run.weco}
\alias{run.weco}
\title{Run Web-Based \code{weco} application}
\usage{
run.weco()
}
\description{
Call Shiny to run \code{weco} as a web-based application. A web browser will
be brought up.
}
\examples{

\dontrun{
run.weco()}


}

