% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weco_rules.R
\name{summary.weco}
\alias{summary.weco}
\title{Summary of WECO outcome}
\usage{
\method{summary}{weco}(object, ...)
}
\arguments{
\item{object}{A class \code{weco} object returned from
\code{\link{weco.rule}} or \code{\link{weco.combine}}}

\item{...}{Other arguments for \code{summary}}
}
\description{
Print summary of a class \code{weco} object
}
\examples{
x        <- rnorm(1000);
rst.comb <- weco.combine(x, lst.rules=list(list(1), list(2, k=9)));
summary(rst.comb);

}

